New GridOrganizer API
The new way to use a Grid is to add all shapes first, then, shuffle the array and set the positions.

This means that the Grid is in charge of setting the positions of the GameShapes. So it is not so much as just a wrapper as a help class.

Removed the old function to plate a shape at a random position.