Fixed the grid resize
Fixed the getSizeRatio function with the proper implementation. The ratio is strictly width to height.

Fixed swapping rows and columns when switching orientation.

Fixed shapes being resized and re-centered on their grid cell. There is a bug where the shapes get de-ranzomized after a resize. For example, when resizing with only 1 shape on the screen. The shape will always to the the grid slot #1 even though it should stay at it’s starting random position. Will investigate.

Changed the ratio of Portrait/Landscape to 1. Anything below 1 is for portrait and higher is for landscape.