Scene size rework
When creating the scene, we check the size of the device screen for its aspect ratio.

I changed the resize algorithm to aspect fill that will resize it. As long as we have the right aspect ratio it will scale up or down without letterboxing.

The base resolutions are separate for iPad and iPhone (square ratios and tall screens) and are still being worked on.