Orientation change handling
Added function to handle orientation change from the Viewcontroller in iOS. This has the view controller call the game scene to resize itself to occupy the new space.

Fixed compilation error on OSX by using the correct method of getting the view size. On Mac OS we don’t go and get the screen size because it’s just for debugging, I’m returning 500x500 for now.