Added Microsoft AppCenter
AppCenter is an analytics suite. It sends data to my account so I can catch crashes and usage data.

Right now it is enabled by default but soon there will be privacy options to disable it. It’s just test versions for now and the app isn’t distributed at the moment.

Now, to build the project, we must use CocoaPods. It is available from homebrew and it is used to download the AppCenter (and others) dependencies. Calling “pod init” will get the necessary files. I wish it was adding it as submodules as I have to ignore the /pods directory and keep regenerating the packages.