Preferences persistance
Added a ViewModel class to pass data back and forth between the view and the game classes.

Added a delegate class to plug in the settings view to receive calls when the view is OK’ed or cancelled.

Still a work in progress, the first Cheats option never seems to save properly for some reason.