Stop scene resizing
This stops the scene from changing size, instead opting to scale down the shapes when they are on the grid.

This is handy because it keeps every element in the scene at a normal size instead of always shrinking down. Bunch of commented out code while I clean everything out.